lift import address data function to context
