+++ src/renderer/contexts/main/DataBackup/types.ts
@@ -1,6 +1,17 @@
 // Copyright 2024 @polkadot-live/polkadot-live-app authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
+export type ImportFunc = (
+  ev: MessageEvent,
+  fromBackup: boolean
+) => Promise<void>;
+
+export type RemoveFunc = (ev: MessageEvent) => Promise<void>;
+
 export interface DataBackupContextInterface {
-  tmp?: null;
+  importAddressData: (
+    serialized: string,
+    handleImportAddress: ImportFunc,
+    handleRemoveAddress: RemoveFunc
+  ) => Promise<void>;
 }
