+++ src/renderer/utils/ImportUtils.ts
@@ -6,7 +6,6 @@ import { Config as ConfigRenderer } from '@/config/processes/renderer';
 import { Flip, toast } from 'react-toastify';
 import type {
   AccountSource,
-  FlattenedAccounts,
   LedgerLocalAddress,
   LocalAddress,
 } from '@/types/accounts';
@@ -166,80 +165,6 @@ export const getSortedLocalLedgerAddresses = (
   }
 
   return sorted;
-};
-
-/**
- * @name importAddresses
- * @summary Extract address data from an imported text file and send to application.
- * (main renderer)
- */
-export const importAddresses = async (
-  serialized: string,
-  handleImportAddress: (ev: MessageEvent, fromBackup: boolean) => Promise<void>,
-  handleRemoveAddress: (ev: MessageEvent) => Promise<void>,
-  setAddresses: (a: FlattenedAccounts) => void
-) => {
-  const s_addresses = getFromBackupFile('addresses', serialized);
-  if (!s_addresses) {
-    return;
-  }
-
-  const p_array: [AccountSource, string][] = JSON.parse(s_addresses);
-  const p_map = new Map<AccountSource, string>(p_array);
-  const importWindowOpen = await window.myAPI.isViewOpen('import');
-
-  for (const [source, ser] of p_map.entries()) {
-    const parsed =
-      source === 'ledger'
-        ? (JSON.parse(ser) as LedgerLocalAddress[])
-        : (JSON.parse(ser) as LocalAddress[]);
-
-    // Check connection status and set isImported to `false` if app is offline.
-    const isOnline: boolean =
-      (await window.myAPI.sendConnectionTaskAsync({
-        action: 'connection:getStatus',
-        data: null,
-      })) || false;
-
-    // Process parsed addresses.
-    for (const a of parsed) {
-      a.isImported && !isOnline && (a.isImported = false);
-
-      // Persist or update address in Electron store.
-      await window.myAPI.rawAccountTask({
-        action: 'raw-account:import',
-        data: { source, serialized: JSON.stringify(a) },
-      });
-
-      // Add address and its status to import window's state.
-      importWindowOpen &&
-        postToImport('import:account:add', { json: JSON.stringify(a), source });
-
-      // Handle importing or removing account from main window and setting `isImported` flag state.
-      const { address, name } = a;
-      const chainId = getAddressChainId(address);
-
-      if (a.isImported) {
-        const data = { data: { data: { address, chainId, name, source } } };
-        await handleImportAddress(new MessageEvent('message', data), true);
-        postToImport('import:address:update', { address: a, source });
-      } else {
-        const data = { data: { data: { address, chainId } } };
-        await handleRemoveAddress(new MessageEvent('message', data));
-        postToImport('import:address:update', { address: a, source });
-      }
-
-      // Update managed account names.
-      const account = AccountsController.get(chainId, address);
-      if (account) {
-        account.name = name;
-        AccountsController.update(chainId, account);
-      }
-    }
-
-    // Update account list state.
-    setAddresses(AccountsController.getAllFlattenedAccountData());
-  }
 };
 
 /**
@@ -407,7 +332,7 @@ export const importAccountSubscriptions = async (
  * @summary Get some serialized data from backup files.
  * Key may be `addresses`, `events` or `intervals`.
  */
-const getFromBackupFile = (
+export const getFromBackupFile = (
   key: string,
   serialized: string
 ): string | undefined => {
@@ -421,6 +346,6 @@ const getFromBackupFile = (
  * @summary Utility to post a message to the import window.
  * (main renderer)
  */
-const postToImport = (task: string, dataObj: AnyData) => {
+export const postToImport = (task: string, dataObj: AnyData) => {
   ConfigRenderer.portToImport?.postMessage({ task, data: dataObj });
 };
